+++ runs/usmarthome/smartbase.py
@@ -1,640 +1,12 @@
 #!/usr/bin/python3
-
-import subprocess
-import json
 import time
 import os
-import urllib.request
+from usmarthome.smartbase0 import Sbase0
+from usmarthome.smartmeas import Slsdm630, Slsdm120, Slwe514, Slfronius
+from usmarthome.smartmeas import Sljson, Slsmaem, Slshelly, Sltasmota, Slmqtt
+from usmarthome.smartmeas import Slhttp, Slavm, Slmystrom
+from usmarthome.smartbut import Sbshelly
 from datetime import datetime, timezone
-
-
-class Sbase0:
-    _basePath = '/var/www/html/openWB'
-    _prefixpy = _basePath+'/modules/smarthome/'
-
-    def logClass(self, level, msg):
-        if (int(level) >= 0):
-            local_time = datetime.now(timezone.utc).astimezone()
-            with open(self._basePath+'/ramdisk/smarthome.log', 'a',
-                      encoding='utf8', buffering=1) as file:
-                if (int(level) == 0):
-                    file.write(local_time.strftime(format="%Y-%m-%d %H:%M:%S")
-                               + '-: ' + str(msg) + '\n')
-                if (int(level) == 1):
-                    file.write(local_time.strftime(format="%Y-%m-%d %H:%M:%S")
-                               + '-: ' + str(msg) + '\n')
-                if (int(level) == 2):
-                    file.write(local_time.strftime(format="%Y-%m-%d %H:%M:%S")
-                               + '-: ' + str(msg) + '\n')
-
-    def readret(self):
-        with open(self._basePath+'/ramdisk/smarthome_device_ret' +
-                  str(self.device_nummer), 'r') as f1:
-            answer = json.loads(json.load(f1))
-        return answer
-
-
-class Slbase(Sbase0):
-    def __init__(self):
-        #
-        # setting
-        print('__init__ Slbase executed')
-        self.device_nummer = 0
-        self.device_name = 'none'
-        self.device_type = 'none'
-        self.temp0 = '300'
-        self.temp1 = '300'
-        self.temp2 = '300'
-        self.devuberschuss = 0
-        self.newwatt = 0
-        self.newwattk = 0
-        self.relais = 0
-        self._smart_param = {}
-        self._device_differentmeasureoment = 0
-        self._device_configured = '0'
-        self._device_ip = 'none'
-        self._device_measuretype = 'none'
-        self._device_measureip = 'none'
-        self._device_measureportsdm = '8899'
-        self._device_measureid = '0'
-        self._device_measuresmaser = '123'
-        self._device_measuresmaage = 15
-        self._device_leistungurl = 'none'
-        self._device_stateurl = 'none'
-        self._device_measureurl = 'none'
-        self._device_measureurlc = 'none'
-        self._device_measurejsonurl = 'none'
-        self._device_measurejsonpower = 'none'
-        self._device_measurejsoncounter = 'none'
-        self._device_measureavmactor = 'none'
-        self._device_measureavmusername = 'none'
-        self._device_measureavmpassword = 'none'
-        self._device_actor = 'none'
-        self._device_username = 'none'
-        self._device_password = 'none'
-
-    def updatepar(self, input_param):
-        self._smart_param = input_param.copy()
-        self.device_nummer = int(self._smart_param.get('device_nummer', '0'))
-        for key, value in self._smart_param.items():
-            try:
-                valueint = int(value)
-            except Exception:
-                valueint = 0
-            # params known to be used in sbase, to avoid logging
-            if (key in ['device_nummer', 'device_mineinschaltdauer',
-                        'device_finishTime', 'device_ausschaltschwelle',
-                        'device_manual_control', 'device_canSwitch',
-                        'device_standbyDuration', 'device_startTime',
-                        'device_onuntilTime', 'device_einschaltverzoegerung',
-                        'device_standbyPower', 'device_einschaltschwelle',
-                        'device_ausschaltverzoegerung',
-                        'device_speichersocbeforestop',
-                        'device_homeConsumtion',
-                        'device_deactivateWhileEvCharging',
-                        'device_startupMulDetection', 'device_onTime',
-                        'device_speichersocbeforestart', 'device_endTime',
-                        'device_maxeinschaltdauer', 'mode',
-                        'WHImported_temp', 'RunningTimeToday',
-                        'oncountnor', 'OnCntStandby', 'device_deactivateper',
-                        'device_startupDetection']):
-                pass
-            elif (key == 'device_differentMeasurement'):
-                self._device_differentmeasurement = valueint
-            elif (key == 'device_type'):
-                self.device_type = value
-            elif (key == 'device_configured'):
-                self._device_configured = value
-            elif (key == 'device_name'):
-                self.device_name = value
-            elif (key == 'device_temperatur_configured'):
-                self.device_temperatur_configured = valueint
-            elif (key == 'device_ip'):
-                self._device_ip = value
-            elif (key == 'device_measureType'):
-                self._device_measuretype = value
-            elif (key == 'device_measureip'):
-                self._device_measureip = value
-            elif (key == 'device_measurePortSdm'):
-                self._device_measureportsdm = value
-            elif (key == 'device_measuresmaage'):
-                self._device_measuresmaage = valueint
-            elif (key == 'device_measuresmaser'):
-                self._device_measuresmaser = value
-            elif (key == 'device_measureid'):
-                self._device_measureid = value
-            elif (key == 'device_leistungurl'):
-                self._device_leistungurl = value
-            elif (key == 'device_measureurl'):
-                self._device_measureurl = value
-            elif (key == 'device_measureurlc'):
-                self._device_measureurlc = value
-            elif (key == 'device_measurejsonurl'):
-                self._device_measurejsonurl = value
-            elif (key == 'device_measurejsonpower'):
-                self._device_measurejsonpower = value
-            elif (key == 'device_measurejsoncounter'):
-                self._device_measurejsoncounter = value
-            elif (key == 'device_measureavmactor'):
-                self._device_measureavmactor = value
-            elif (key == 'device_measureavmusername'):
-                self._device_measureavmusername = value
-            elif (key == 'device_measureavmpassword'):
-                self._device_measureavmpassword = value
-            elif (key == 'device_actor'):
-                self._device_actor = value
-            elif (key == 'device_username'):
-                self._device_username = value
-            elif (key == 'device_password'):
-                self._device_password = value
-            elif (key == 'device_stateurl'):
-                self._device_stateurl = value
-            else:
-                self.logClass(2, "(" + str(self.device_nummer) + ") "
-                              + __class__.__name__ + " überlesen " + key +
-                              " " + value)
-
-    def __del__(self):
-        print('__del__ Slbase executed ')
-
-
-class Slmqtt(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slmqtt excuted')
-
-    def getwattread(self):
-        self._watt(self._device_ip)
-
-    def sepwattread(self):
-        self._watt(self._device_measureip)
-        return self.newwatt, self.newwattk
-
-    def _watt(self, ip):
-        argumentList = ['python3', self._prefixpy + 'mqtt/watt.py',
-                        str(self.device_nummer), str(ip),
-                        str(self.devuberschuss)]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-            self.relais = int(answer['on'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('mqtt', self.device_nummer, ip, str(e1)))
-
-
-class Slshelly(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slshelly excuted')
-
-    def getwattread(self):
-        self._watt(self._device_ip)
-
-    def sepwattread(self):
-        self._watt(self._device_measureip)
-        return self.newwatt, self.newwattk
-
-    def _watt(self, ip):
-        argumentList = ['python3', self._prefixpy + 'shelly/watt.py',
-                        str(self.device_nummer), str(ip), '0']
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-            self.relais = int(answer['on'])
-            if (self.device_temperatur_configured > 0):
-                self.temp0 = str(answer['temp0'])
-                with open(self._basePath+'/ramdisk/device' +
-                          str(self.device_nummer) + '_temp0', 'w') as f:
-                    f.write(str(self.temp0))
-            else:
-                self.temp0 = '300'
-            if (self.device_temperatur_configured > 1):
-                self.temp1 = str(answer['temp1'])
-                with open(self._basePath+'/ramdisk/device' +
-                          str(self.device_nummer) + '_temp1', 'w') as f:
-                    f.write(str(self.temp1))
-            else:
-                self.temp1 = '300'
-            if (self.device_temperatur_configured > 2):
-                self.temp2 = str(answer['temp2'])
-                with open(self._basePath+'/ramdisk/device' +
-                          str(self.device_nummer) + '_temp2', 'w') as f:
-                    f.write(str(self.temp2))
-            else:
-                self.temp2 = '300'
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('Shelly', self.device_nummer, ip, str(e1)))
-
-
-class Slavm(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slavm excuted')
-
-    def getwattread(self):
-        self._watt(self._device_ip, self._device_actor,
-                   self._device_username,
-                   self._device_password)
-
-    def sepwattread(self):
-        self._watt(self._device_measureip,
-                   self._device_measureavmactor,
-                   self._device_measureavmusername,
-                   self._device_measureavmpassword)
-        return self.newwatt, self.newwattk
-
-    def _watt(self, ip, act, user, pw):
-        argumentList = ['python3', self._prefixpy +
-                        'avmhomeautomation/watt.py',
-                        str(self.device_nummer), str(ip),
-                        '0', '0',
-                        act, user, pw]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-            self.relais = int(answer['on'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('Avm ', self.device_nummer, ip, str(e1)))
-
-
-class Sltasmota(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Sltasmota excuted')
-
-    def getwattread(self):
-        self._watt(self._device_ip)
-
-    def sepwattread(self):
-        self._watt(self._device_measureip)
-        return self.newwatt, self.newwattk
-
-    def _watt(self, ip):
-        argumentList = ['python3', self._prefixpy + 'tasmota/watt.py',
-                        str(self.device_nummer), str(ip), '0']
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-            self.relais = int(answer['on'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('Tasmota', self.device_nummer, ip, str(e1)))
-
-
-class Slhttp(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slhttp excuted')
-
-    def getwattread(self):
-        self._watt(self._device_leistungurl, 'none',
-                   self._device_stateurl)
-
-    def sepwattread(self):
-        self._watt(self._device_measureurl, self._device_measureurlc,
-                   'none')
-        return self.newwatt, self.newwattk
-
-    def _watt(self, url, urlc, urls):
-        argumentList = ['python3', self._prefixpy + 'http/watt.py',
-                        str(self.device_nummer), '0',
-                        str(self.devuberschuss), url, urlc,
-                        '0', '0', urls]
-        proc = subprocess.Popen(argumentList)
-        proc.communicate()
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-            self.relais = int(answer['on'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d Fehlermeldung: %s "
-                          % ('http', self.device_nummer, str(e1)))
-
-
-class Slmystrom(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slmystrom excuted')
-
-    def getwattread(self):
-        self._watt(self._device_ip)
-
-    def sepwattread(self):
-        self._watt(self._device_measureip)
-        return self.newwatt, self.newwattk
-
-    def _watt(self, ip):
-        argumentList = ['python3', self._prefixpy + 'mystrom/watt.py',
-                        str(self.device_nummer), str(ip), '0']
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-            self.relais = int(answer['on'])
-            if (self.device_temperatur_configured > 0):
-                self.temp0 = str(answer['temp0'])
-                with open(self._basePath+'/ramdisk/device' +
-                          str(self.device_nummer) + '_temp0', 'w') as f:
-                    f.write(str(self.temp0))
-            else:
-                self.temp0 = '300'
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('Mystrom', self.device_nummer, ip, str(e1)))
-
-
-class Slsmaem(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slsmaem excuted')
-
-    def sepwattread(self):
-        argumentList = ['python3', self._prefixpy + 'smaem/watt.py',
-                        str(self.device_nummer), str(self._device_measureip),
-                        str(self._device_measuresmaser),
-                        str(self._device_measuresmaage)]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('smaem ', self.device_nummer,
-                             str(self._device_measureip), str(e1)))
-        return self.newwatt, self.newwattk
-
-
-class Slwe514(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slwe514 excuted')
-
-    def sepwattread(self):
-        argumentList = ['python3', self._prefixpy + 'we514/watt.py',
-                        str(self.device_nummer), str(self._device_measureip),
-                        str(self._device_measureid)]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('we514 ', self.device_nummer,
-                             str(self._device_measureip), str(e1)))
-        return self.newwatt, self.newwattk
-
-
-class Sljson(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Sljson excuted')
-
-    def sepwattread(self):
-        argumentList = ['python3', self._prefixpy + 'json/watt.py',
-                        str(self.device_nummer),
-                        self._device_measurejsonurl,
-                        self._device_measurejsonpower,
-                        self._device_measurejsoncounter]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %s Fehlermeldung: %s "
-                          % ('json ', self.device_nummer, str(e1)))
-        return self.newwatt, self.newwattk
-
-
-class Slfronius(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slfronius excuted')
-
-    def sepwattread(self):
-        argumentList = ['python3', self._prefixpy + 'fronius/watt.py',
-                        str(self.device_nummer), str(self._device_measureip),
-                        str(self._device_measureid)]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('fronius ', self.device_nummer,
-                             str(self._device_measureip), str(e1)))
-        return self.newwatt, self.newwattk
-
-
-class Slsdm630(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slsdm630 excuted')
-
-    def sepwattread(self):
-        argumentList = ['python3', self._prefixpy + 'sdm630/sdm630.py',
-                        str(self.device_nummer), str(self._device_measureip),
-                        str(self._device_measureid),
-                        str(self._device_measureportsdm)]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('Sdm630 ', self.device_nummer,
-                             str(self._device_measureip), str(e1)))
-        return self.newwatt, self.newwattk
-
-
-class Slsdm120(Slbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Slsdm120 excuted')
-
-    def sepwattread(self):
-        argumentList = ['python3', self._prefixpy + 'sdm120/sdm120.py',
-                        str(self.device_nummer), str(self._device_measureip),
-                        str(self._device_measureid),
-                        str(self._device_measureportsdm)]
-        try:
-            proc = subprocess.Popen(argumentList)
-            proc.communicate()
-            answer = self.readret()
-            self.newwatt = int(answer['power'])
-            self.newwattk = int(answer['powerc'])
-        except Exception as e1:
-            self.logClass(2, "Leistungsmessung %s %d %s Fehlermeldung: %s "
-                          % ('Sdm120 ', self.device_nummer,
-                             str(self._device_measureip), str(e1)))
-        return self.newwatt, self.newwattk
-
-
-class Spbase(Sbase0):
-    def __init__(self):
-        #
-        # setting
-        print('__init__ Spbase executed')
-        self._device_pbip = 'none'
-        self.device_nummer = 0
-
-    def updatepar(self, input_param):
-        self._smart_param = input_param.copy()
-        self.device_nummer = int(self._smart_param.get('device_nummer', '0'))
-        for key, value in self._smart_param.items():
-            if (key == 'device_pbip'):
-                self._device_pbip = value
-            else:
-                self.logClass(2, "(" + str(self.device_nummer) + ") "
-                              + __class__.__name__ + " überlesen " + key +
-                              " " + value)
-
-    def showstat(self, manual, relais):
-        pass
-
-
-class Sbshelly(Spbase):
-    def __init__(self):
-        # setting
-        super().__init__()
-        print('__init__ Sbshelly excuted')
-        self.counter = 0
-        self.led = 9
-        self.event_cnt = 0
-        self.event = 'none'
-        self.oldevent_cnt = 0
-        self.oldevent = 'none'
-
-    def showstat(self, manual, relais):
-        if (manual == 0):  # automatic mode
-            self.pboff()
-        else:   # manual mode
-            if (relais == 0):
-                self.pbon()
-            else:
-                self.pbblink()
-
-    def checkbut(self, manual, relais, manual_control):
-        newmanual = manual
-        newmanual_control = manual_control
-        at = str(urllib.request.urlopen("http://" +
-                                        str(self._device_pbip)
-                                        + "/status",
-                                        timeout=3).read().decode("utf-8"))
-        a = json.loads(at)
-        with open(self._basePath+'/ramdisk/smarthome_device_ret' +
-                  str(self.device_nummer) + '_shelly_bp', 'w') as f:
-            f.write(str(a))
-        self.oldevent_cnt = self.event_cnt
-        self.oldevent = self.event
-        self.event_cnt = int(a['inputs'][0]['event_cnt'])
-        self.event = str(a['inputs'][0]['event'])
-        if (self.oldevent == 'none'):
-            return newmanual, newmanual_control
-        if ((self.event == self.oldevent) and
-           (self.event_cnt == self.oldevent_cnt)):
-            return newmanual, newmanual_control
-        self.logClass(2, "Shelly button pressed (%d) %s %s"
-                      % (self.device_nummer,
-                         self._device_pbip, self.event))
-        # im automatic modus -> ein mal Drücken wechselen auf manual
-        if (manual == 0):
-            newmanual = 1
-            return newmanual, newmanual_control
-        # im manual modus  -> ein mal Drücken wechselen  zwischen on und off
-        if (self.event == 'S'):
-            if (manual_control == 1):
-                newmanual_control = 0
-            else:
-                newmanual_control = 1
-            return newmanual, newmanual_control
-        # im manual modus  -> mehrmals drücken wechselen auf automat
-        newmanual = 0
-        return newmanual, newmanual_control
-
-    def pboff(self):
-        if (self.led == 0):
-            return
-        try:
-            urllib.request.urlopen("http://" + str(self._device_pbip) +
-                                   "/settings?led_status_disable=true",
-                                   timeout=3)
-            self.logClass(2, "Shelly button led off (%d) %s"
-                          % (self.device_nummer,
-                             self._device_pbip))
-
-        except Exception as e1:
-            self.logClass(2, "Shelly button off (%d) %s Fehlermeldung: %s "
-                          % (self.device_nummer,
-                             self._device_pbip, str(e1)))
-        self.led = 0
-
-    def pbon(self):
-        if (self.led == 1):
-            return
-        try:
-            urllib.request.urlopen("http://" + str(self._device_pbip) +
-                                   "/settings?led_status_disable=false",
-                                   timeout=3)
-            self.logClass(2, "Shelly button led on (%d) %s"
-                          % (self.device_nummer,
-                             self._device_pbip))
-        except Exception as e1:
-            self.logClass(2, "Shelly button on (%d) %s Fehlermeldung: %s "
-                          % (self.device_nummer,
-                             self._device_pbip, str(e1)))
-        self.led = 1
-
-    def pbblink(self):
-        self.counter = self.counter + 1
-        if (self.counter < 1):
-            return
-        self.counter = 0
-        if (self.led == 0):
-            self.pbon()
-        else:
-            self.pboff()
 
 
 class Sbase(Sbase0):
